+++ src/unplugin.ts
@@ -50,7 +50,7 @@ export default createUnplugin<Partial<UnimportPluginOptions>>((options = {}) => 
       }
 
       if (dts) {
-        return fs.writeFile(dts, ctx.generateTypeDecarations(), 'utf-8')
+        return fs.writeFile(dts, ctx.generateTypeDeclarations(), 'utf-8')
       }
     }
   }